Changed double to decimal; fixed problem with move date changing test result; deactivated schedules that was active with account closed; changed yellow and red limit to cents; task list review.
by [Dara Keon (desktop)] in [2015-01-03 11:27:10 (-02:00)] - hg hash {c293053d1c90}

+++ site/DFM.BusinessLogic/Repositories/MonthRepository.cs
@@ -59,12 +59,12 @@ namespace DFM.BusinessLogic.Repositories
 
 
 
-		internal Double GetIn(Year year, Category category)
+		internal Decimal GetIn(Year year, Category category)
 		{
 			return get(year, category).Sum(m => m.In);
 		}
 
-	    internal Double GetOut(Year year, Category category)
+        internal Decimal GetOut(Year year, Category category)
 		{
 			return get(year, category).Sum(m => m.Out);
 		}
