+++ site/DFM.Generic/NumericExtension.cs
@@ -4,28 +4,28 @@ namespace DFM.Generic
 {
 	public static class NumericExtension
 	{
-		public static Int32 ToCents(this Double value)
+        public static Int32 ToCents(this Decimal value)
 		{
             return (Int32)(Math.Round(value * 100, 0));
 		}
 
-		public static Double ToVisual(this Int32 value)
+        public static Decimal ToVisual(this Int32 value)
 		{
-			return value / 100.0;
+			return value / 100m;
 		}
 
-		public static Int32? ToCents(this Double? value)
+        public static Int32? ToCents(this Decimal? value)
 		{
 		    return value.HasValue 
                 ? value.Value.ToCents()
                 : default(Int32?);
 		}
 
-		public static Double? ToVisual(this Int32? value)
+        public static Decimal? ToVisual(this Int32? value)
 		{
 			return value.HasValue
                 ? value.Value.ToVisual()
-                : default(Double?);
+                : default(Decimal?);
 		}
 
 	}
